Rename platformProvider
